Match Geography URI

Matches passed string with URI available in the Standard Geography Code Register.For URI's function returned passed values.

find_geography_URI(geography, database = "internal")

Arguments

geography
A character vector of geography names. URIs starting with http are returned as such without searching the data set.

database
Can be one of the following internal, ScotStat, both. Defaults to internal that reflects an extract of Standard Geography Code Register. Option online searches across the Statistics.gov.scot.

Value

A named character vector of unique geography URIs. For multiple matches names will be numbered - string 'Edinburgh' will match multiple geographies that will be named 'Edinburgh1', 'Edinburgh2', ...

Details

The function is mostly used internally to make using get_geography_data more user friendly. The function can be used in isolation to conveniently get URIs for the colloquial geography names.

Examples

find_geography_URI(geography = c("Edinburgh", "Glasgow"))

#> Geography 'Edinburgh' matched multiple URIs

#> Geography 'Glasgow' matched multiple URIs

#>                                                     Glasgow.NA1 
#> "http://statistics.gov.scot/id/statistical-geography/S19000510" 
#>                                                   Edinburgh.NA1 
#> "http://statistics.gov.scot/id/statistical-geography/S19000550" 
#>                                                   Edinburgh.NA2 
#> "http://statistics.gov.scot/id/statistical-geography/S19000855" 
#>                                                     Glasgow.NA3 
#> "http://statistics.gov.scot/id/statistical-geography/S19000899" 
#>                                                   Edinburgh.NA6 
#> "http://statistics.gov.scot/id/statistical-geography/S19001502" 
#>                                                     Glasgow.NA2 
#> "http://statistics.gov.scot/id/statistical-geography/S19001549" 
#>                                                   Edinburgh.NA3 
#> "http://statistics.gov.scot/id/statistical-geography/S20000463" 
#>                                                   Edinburgh.NA9 
#> "http://statistics.gov.scot/id/statistical-geography/S20000682" 
#>                                                   Edinburgh.NA8 
#> "http://statistics.gov.scot/id/statistical-geography/S20001198" 
#>                                                   Edinburgh.NA5 
#> "http://statistics.gov.scot/id/statistical-geography/S22000012" 
#>                                                     Glasgow.NA4 
#> "http://statistics.gov.scot/id/statistical-geography/S22000018" 
#>                                                   Edinburgh.NA7 
#> "http://statistics.gov.scot/id/statistical-geography/S22000059" 
#>                                                     Glasgow.NA5 
#> "http://statistics.gov.scot/id/statistical-geography/S22000065" 
#>                                                   Edinburgh.NA4 
#> "http://statistics.gov.scot/id/statistical-geography/S37000012"